Move project to root module
The `Project` struct has been moved from the `cli` module to the root
of the crate. This makes it easier to reuse the struct and its
associated testing code across the crate.
