+++ src/cli/commands/mod.rs
@@ -2,7 +2,7 @@ use anyhow::Error;
 use async_trait::async_trait;
 use clap::Subcommand;
 
-use crate::cli::project::Project;
+use crate::Project;
 
 pub use self::create::Create;
 pub use self::init::Init;
