+++ src/lib.rs
@@ -1,4 +1,4 @@
-pub use self::{error::*, fragment::*, renderer::*, template::*, workflow::*};
+pub use self::{error::*, fragment::*, project::*, renderer::*, template::*, workflow::*};
 
 #[cfg(feature = "cli")]
 pub mod cli;
@@ -6,6 +6,7 @@ pub mod cli;
 mod error;
 mod fragment;
 pub mod github;
+mod project;
 mod renderer;
 mod template;
 mod workflow;
